refactor(Endpoint): Split files into dedicated extensiions feat(Endpoint): Finish the external links endpoint implementation
